Fixed a problem with dlType in DTO
When fixing issue #537, unit tests didn't exercise DTOs and the one
regarding the dlType failed in the functional test.

This is a fast fix which forces the user of setDlType to use Integers
instead of Shorts. The problem was that when converting from Short to
Integer in that function, Java complained about precision loss.

Reference: issue #537

+++ midonet-client/src/main/java/org/midonet/client/dto/DtoRule.java
@@ -197,7 +197,7 @@ public class DtoRule {
         return dlType;
     }
 
-    public void setDlType(Short dlType) {
+    public void setDlType(Integer dlType) {
         Integer intDlType = null;
         if (dlType != null) {
             intDlType = dlType & 0xffff;
